style: SideBar 컴포넌트 구현 - 캘린더 리스트 CSS 수정 (#28)
