--- front/src/components/SideBar/SubscriptionCalendar/style.module.css
@@ -1,22 +1,0 @@
-.calendar_details{
-    width: 230px;
-}
-
-.calendar_details > * {
-    display: flex;
-    align-items: center;
-    padding: 0px 15px;
-}
-
-.calendar_details > summary {
-    height: 30px;
-    list-style: none;
-    
-    justify-content: space-between;
-    font-size: 14px;
-    margin: 10px 0px;
-}
-
-.calendar_details > summary:hover{
-    background-color: #f8f8f8;
-}
