+++ front/src/components/SideBar/calendars/CalendarInfo/index.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import styles from './style.module.css';
-import PropTypes from 'props-type';
+import styles_calendar from './style.module.css';
+import PropTypes from 'prop-types';
 
 const Index = ({ calendars }) => {
   return (
     <>
       {calendars.map(({ id, calendarName }) => (
-        <label key={id} className={styles.calendarInfo_label}>
+        <label key={id} className={styles_calendar.calendarInfo_label}>
           <input type="checkbox" /> {calendarName}
         </label>
       ))}
@@ -15,7 +15,7 @@ const Index = ({ calendars }) => {
 };
 
 Index.propTypes = {
-  calendars: PropTypes.object,
+  calendars: PropTypes.array,
 };
 
 export default Index;
