+++ front/src/components/SideBar/calendars/style.module.css
@@ -1,11 +1,7 @@
-.calendar_details{
-    width: 230px;
-}
-
 .calendar_details > * {
     display: flex;
     align-items: center;
-    padding: 0px 15px;
+    padding: 0px 20px;
 }
 
 .calendar_details > summary {
